Pull data directly from Overpass based on bounding box of map.
Only loads data for currently active layer.
Is slightly slower since we are waiting on load as we pan.

Only request and setData when requested bounds are beyond previous
so zooming in then back out does not prompt an API call.

We request data from overpass for a slightly larger box than the one
the map currently displays.  This is done because objects just off-
screen may have a fill that extends on-screen.